model-server/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
torch==2.2.0
transformers==4.37.2
sentence-transformers==2.3.1
pydantic==2.6.1
python-multipart==0.0.9
numpy==1.26.3
scikit-learn==1.4.0
datasets==2.16.1
accelerate==0.26.1
peft==0.8.2
trl==0.7.11
bitsandbytes==0.42.0
sse-starlette==1.8.2
pytest==8.0.0
pytest-asyncio==0.23.4
faiss-cpu==1.7.4
redis==5.0.1
PyYAML==6.0.1
